+++ runners/s3-benchrunner-c/CS3Runner.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include "BenchmarkRunner.h"
+
+// Create runner that uses aws-c-s3 directly
+std::unique_ptr<BenchmarkRunner> createCS3BenchmarkRunner(const BenchmarkConfig &config);
