+++ runners/s3-benchrunner-cpp/SdkTransferManagerRunner.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include "BenchmarkRunner.h"
+
+// Create runner that uses C++ SDK's transfer manager
+std::unique_ptr<BenchmarkRunner> createSdkTransferManagerRunner(const BenchmarkConfig &config);
