+++ scripts/utils/__init__.py
@@ -35,6 +35,7 @@ def _add_runner(runner: Runner):
 
 
 _add_runner(Runner('c', s3_clients=['crt-c']))
+_add_runner(Runner('cpp', s3_clients=['sdk-cpp-client-crt', 'sdk-cpp-client-classic', 'sdk-cpp-tm-classic']))
 _add_runner(Runner('java',
                    s3_clients=['crt-java', 'sdk-java-client-crt', 'sdk-java-tm-crt']))
 _add_runner(Runner('python',
@@ -138,7 +139,7 @@ def fetch_git_repo(url: str, dir: Path, main_branch: str = 'main', preferred_bra
         run(['git', 'pull'])
 
     # update submodules
-    run(['git', 'submodule', 'update', '--init'])
+    run(['git', 'submodule', 'update', '--init', '--recursive'])
 
     os.chdir(cwd_prev)
