+++ scripts/utils/build.py
@@ -10,7 +10,7 @@ from typing import Optional
 from utils import fetch_git_repo, run, RUNNERS
 
 
-def _build_cmake_proj(src_dir: Path, build_dir: Path, install_dir: Path):
+def _build_cmake_proj(src_dir: Path, build_dir: Path, install_dir: Path, config_extra: list[str] = []):
 
     config_cmd = ['cmake',
                   '-S', str(src_dir),
@@ -35,6 +35,8 @@ def _build_cmake_proj(src_dir: Path, build_dir: Path, install_dir: Path):
     if src_dir.name != 's3-benchrunner-c':
         # runner doesn't have tests
         config_cmd += ['-DBUILD_TESTING=OFF']
+
+    config_cmd += config_extra
 
     run(config_cmd)
     run(build_cmd)
@@ -79,6 +81,32 @@ def _build_c(work_dir: Path, branch: Optional[str]) -> list[str]:
 
     # return runner cmd
     return [str(install_dir/'bin/s3-benchrunner-c')]
+
+
+def _build_cpp(work_dir: Path, branch: Optional[str]) -> list[str]:
+    """build s3-benchrunner-cpp"""
+
+    install_dir = work_dir/'install'
+
+    # build C++ SDK
+    fetch_git_repo(url='https://github.com/aws/aws-sdk-cpp.git',
+                   dir=work_dir/'aws-sdk-cpp',
+                   preferred_branch=branch)
+    _build_cmake_proj(
+        src_dir=work_dir/'aws-sdk-cpp',
+        build_dir=work_dir/'build',
+        install_dir=install_dir,
+        config_extra=['-DBUILD_SHARED_LIBS=OFF',
+                      '-DENABLE_TESTING=OFF',
+                      '-DBUILD_ONLY=s3-crt;transfer'])
+
+    # build s3-benchrunner-cpp
+    _build_cmake_proj(src_dir=RUNNERS['cpp'].dir,
+                      build_dir=work_dir/'s3-benchrunner-cpp-build',
+                      install_dir=install_dir)
+
+    # return runner cmd
+    return [str(install_dir/'bin/s3-benchrunner-cpp')]
 
 
 def _fetch_and_install_python_repo(
@@ -209,6 +237,7 @@ def build_runner(lang: str, build_root_dir: Path, branch: Optional[str]) -> list
     # get build function by name, and call it
     build_functions = {
         'c': _build_c,
+        'cpp': _build_cpp,
         'python': _build_python,
         'java': _build_java,
     }
